% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA4GHshiny-package.R
\docType{package}
\name{GA4GHshiny-package}
\alias{GA4GHshiny-package}
\alias{GA4GHshiny}
\title{Shiny application for interacting with GA4GH-based data servers}
\description{
GA4GHshiny package provides an easy way to interact with data servers based
on Global Alliance for Genomics and Health (GA4GH) Genomics API through a
Shiny application. It also integrates with Beacon Network.
}
\author{
Welliton Souza, Benilton Carvalho, Cristiane Rocha, Elizabeth Borgognoni

Maintainer: Welliton Souza <well309@gmail.com>
}
\keyword{package}
